refactor: move bond nominate modals to default and shards subfolders
