+++ src/renderer/widgets/Staking/BondNominate/default/index.ts
