+++ src/renderer/widgets/Staking/BondNominate/default/lib/bond-utils.ts
