+++ src/renderer/widgets/Staking/BondNominate/default/lib/types.ts
