+++ src/renderer/widgets/Staking/BondNominate/default/model/bond-nominate-model.ts
