+++ src/renderer/widgets/Staking/BondNominate/default/ui/BondForm.tsx
