+++ src/renderer/widgets/Staking/BondNominate/default/ui/BondNominate.tsx
