+++ src/renderer/widgets/Staking/BondNominate/shards/index.ts
