+++ src/renderer/widgets/Staking/BondNominate/shards/lib/bond-utils.ts
