+++ src/renderer/widgets/Staking/BondNominate/shards/lib/types.ts
