+++ src/renderer/widgets/Staking/BondNominate/shards/model/bond-nominate-model.ts
