+++ src/renderer/widgets/Staking/BondNominate/shards/model/form-model.ts
