+++ src/renderer/widgets/Staking/BondNominate/shards/ui/BondForm.tsx
