+++ src/renderer/widgets/Staking/BondNominate/shards/ui/BondNominate.tsx
